Extract common code into a BuildsQueries trait
